renaming of Portuguese heavy commenting Added Crocodile enemy
