+++ rust_hero_data/src/data/locations.rs
@@ -1,11 +1,17 @@
 //! Locations - Data and structs related to locations.
 use crate::data::enemies::Enemies;
 
+/// The locations we can travel to in a region
 pub enum Locations {
+    /// The place to rest and purchase things
     Town,
+    /// Dense undergrowth has lead to many creatures living here
     Forest,
+    /// Inaccessable to many, the cave is full of lurking monsters
     Cave,
+    /// 
     Swamp,
+    /// 
     Desert,
 }
 
@@ -15,6 +21,7 @@ pub struct Location {
 }
 
 impl Location {
+    /// Get a `Location` struct from a `Locations` enum
     pub const fn get_location(location: &Locations) -> Self {
         match location {
             Locations::Town => Self {
@@ -31,7 +38,7 @@ impl Location {
             },
             Locations::Swamp => Self {
                 name: "Swamp",
-                enemies: Some([Enemies::Rat, Enemies::Rabbit, Enemies::Snake, Enemies::Wolf]),
+                enemies: Some([Enemies::Rat, Enemies::Rabbit, Enemies::Snake, Enemies::Crocodile]),
             },
             Locations::Desert => Self {
                 name: "Desert",
